requirements.txt
pymupdf
langchain
langchain-community
faiss-cpu
transformers
sentence-transformers
openai
python-dotenv
